Introduce in-use analysis for speculation
This analysis identifies values that are transitively used in control flow or
in memref subscript computations. Operations defining such values must be
replicated in the speculative function.

The idea of the analysis, but not the implementation, is derived from
differential activity analysis in Enzyme-MLIR. See Peng, Moses, Zinenko and
Dubach, to appear in OOPSLA 2025.

Signed-off-by: Alex Zinenko <git@ozinenko.com>