MLIR op definitions for MoE-relevant kernels
Introduce a small subset of op definitions sufficient to define
MoE-relevant kernels currently expressed in Python.

Introduce an address space to the tensor type and an ODS type constraint
to restrict the introduced arithmetic ops to in-registers tensors and
the memory access ops to the in-memory/register tensors where relevant.

Signed-off-by: Alex Zinenko <git@ozinenko.com>