Simplify map substitution logic
Provide a utility function to directly substitute named symbols from a
hyperparam attribute instead of having to construct an intermediary
value vector. This shortens the code and will come handy in lowerings
were only a subset of symbols need to be substituted, e.g., workgroup
indexes are present but not used in step/size maps.

Signed-off-by: Alex Zinenko <git@ozinenko.com>