Allow Wave ops to use vector types directly (#55)
Completely overhaul the lowering procedure for register-resident Wave
dialect tensors. They are representing values that are virtually
distributed across multiple threads. When lowering to core MLIR
dialects, we also perform a programming model change: the lowered IR is
expected to be SIMT. Therefore, register-resident Wave dialect tensors
should be converted to vectors of a _distributed_ shape rather than a
shape that comes from direct substitution of symbolic values from
hyperparameters (unlike memory-resident tensors).

The distributed per-thread shape of the vector is derived from the
"elements per thread" property available on memory operations, in
hyperparameters and compilation options, or extracted from intrinsic
structure of the MMA operation. Implement a dataflow analysis for
propagating this value from the property for now and report eventual
conflicts.

Based on the above dataflow analysis, implement a pass rewriting
register-resident Wave dialect tensors to vectors with distributed
shapes. Note that this _cannot_ be implemented as a dialect conversion
(or at least is significantly complicated) since the type conversion
depends on the result of the analysis, and not only on the original type
or defining operation.

This in turn requires extending Wave dialect operations to also accept
vector-typed values whenever a register-resident tensor is accepted.
This is not a significant departure from the original representation as
both tensors and vectors are immutable SSA values.

Introduce an additional normalform checking for the absence of
register-resident tensors in the input IR and make it a precondition of
the lowering pass, and a postcondition of the elements-per-thread
propagation pass.

---------

Signed-off-by: Alex Zinenko <git@ozinenko.com>
Signed-off-by: tyb0807 <sontuan.vu@amd.com>
Co-authored-by: tyb0807 <sontuan.vu@amd.com>